[TIMOB-18078] Bug Fix

+++ iphone/Classes/StreamModule.m
@@ -182,8 +182,8 @@
     // Handle asynch
     if (callback != nil) {
         SEL operation = @selector(readToBuffer:offset:length:callback:);
-        int offset = 0;
-        int length = 0;
+        NSInteger offset = 0;
+        NSInteger length = 0;
         
         NSInvocation* invoke = [NSInvocation invocationWithMethodSignature:[stream methodSignatureForSelector:operation]];
         [invoke setTarget:stream];
@@ -226,7 +226,7 @@
                     location:CODELOCATION];        
     }   
     
-    int size = [TiUtils intValue:chunkSize];
+    NSInteger size = [TiUtils intValue:chunkSize];
     if (callback != nil) {
         NSInvocation* invoke = [NSInvocation invocationWithMethodSignature:[inputStream methodSignatureForSelector:@selector(writeToStream:chunkSize:callback:)]];
         [invoke setTarget:inputStream];
@@ -262,7 +262,7 @@
                     location:CODELOCATION];        
     }
     
-    int size = [TiUtils intValue:chunkSize];
+    NSInteger size = [TiUtils intValue:chunkSize];
     BOOL isAsynch = [TiUtils boolValue:asynch def:NO];
     if (isAsynch) {
         NSInvocation* invoke = [NSInvocation invocationWithMethodSignature:[stream methodSignatureForSelector:@selector(pumpToCallback:chunkSize:asynch:)]];
